[Test] Refactor Assert, ApiClient, Context into headers
