+++ server/tests/FullTest.cc
@@ -3,16 +3,12 @@
 #define CONFIG_ALLOW_MUTATION
 #include "Config.hh"
 
-#include "TokenPayload.hh"
 #include "TokenUtils.hh"
 #include "component/AppComponent.hh"
 #include "controller/AuthController.hh"
 #include "controller/OfferController.hh"
 #include "controller/PlaceController.hh"
 #include "controller/UserController.hh"
-#include "dto/OfferDto.hh"
-#include "dto/PlaceDto.hh"
-#include "dto/UserDto.hh"
 
 #include <oatpp-test/UnitTest.hpp>
 #include <oatpp/core/macro/codegen.hpp>
@@ -25,124 +21,11 @@
 #include <oatpp/web/client/HttpRequestExecutor.hpp>
 #include <oatpp/web/protocol/http/incoming/Response.hpp>
 
-#include <memory>
-#include <source_location>
+#include "Assert.hh"
+#include "ApiClient.hh"
+#include "Context.hh"
 
-namespace {
-
-using AssertMessageT = std::function<std::ostream&(std::ostream&)>;
-/*
-  Asserts supplied value and yells at you if it's false.
-
-  When @p value is false, prints detailed location of the function call, prints
-  the message supplied in @p message if it's present, and terminates the
-  program.
-
-  @param message Supposed to be obtained from calling assertWrap().
-
-  @note I prefer not to use OATPP_ASSERT(), because it does not give any usefull
-  feedback on where it failed.
-*/
-void
-testAssert(bool value,
-           std::optional<AssertMessageT> message = {},
-           std::source_location const loc = std::source_location::current())
-{
-    if (!value) {
-        std::cerr << "ASSERT FAILED: " << loc.file_name() << ":" << loc.line()
-                  << ":" << loc.column() << " " << loc.function_name() << '\n';
-        if (message) {
-            (*message)(std::cerr) << '\n';
-        }
-        exit(1);
-        // throw std::runtime_error{"ASSERT FAILED"};
-    }
-}
-
-/*
-  Wrap response into object which can be printed by testAssert.
-*/
-auto
-assertWrap(
-  std::shared_ptr<oatpp::web::protocol::http::incoming::Response> response)
-{
-    // Important is that we execute readBodyToString() *only* inside
-    // testAssert(). That is why AssertMessageT is supposed to be a function.
-    return [=](std::ostream& ostream) -> std::ostream& {
-        ostream << *response->readBodyToString();
-        return ostream;
-    };
-}
-
-#include OATPP_CODEGEN_BEGIN(ApiClient)
-
-class Client : public oatpp::web::client::ApiClient
-{
-  public:
-    API_CLIENT_INIT(Client)
-
-    // User
-    API_CALL("POST",
-             "signup",
-             signup,
-             BODY_DTO(oatpp::Object<server::dto::SignUpDto>, dto))
-    API_CALL("POST",
-             "signin",
-             signin,
-             BODY_DTO(oatpp::Object<server::dto::SignInDto>, dto))
-
-    // Place
-    API_CALL("POST",
-             "place",
-             place_create_one,
-             AUTHORIZATION(String, auth_string, "Bearer"),
-             BODY_DTO(oatpp::Object<server::dto::PlaceDto>, dto))
-    API_CALL("DELETE",
-             "place/{id}",
-             place_delete_one,
-             AUTHORIZATION(String, auth_string, "Bearer"),
-             PATH(UInt64, id))
-
-    // Offer
-    API_CALL("POST",
-             "offer",
-             offer_create_one,
-             AUTHORIZATION(String, auth_string, "Bearer"),
-             BODY_DTO(oatpp::Object<server::dto::OfferDto>, dto))
-    API_CALL("DELETE",
-             "offer/{id}",
-             offer_delete_one,
-             AUTHORIZATION(String, auth_string, "Bearer"),
-             PATH(UInt64, id))
-};
-
-#include OATPP_CODEGEN_END(ApiClient)
-
-struct TestEnvironment
-{
-    TestEnvironment()
-      : conn_provider(
-          oatpp::network::tcp::client::ConnectionProvider::createShared(
-            { "localhost", 8000 }))
-      , executor(
-          oatpp::web::client::HttpRequestExecutor::createShared(conn_provider))
-      , mapper(oatpp::parser::json::mapping::ObjectMapper::createShared())
-      , client(Client::createShared(executor, mapper))
-    {
-    }
-
-    std::shared_ptr<oatpp::network::tcp::client::ConnectionProvider>
-      conn_provider;
-    std::shared_ptr<oatpp::web::client::HttpRequestExecutor> executor;
-    std::shared_ptr<oatpp::parser::json::mapping::ObjectMapper> mapper;
-    std::shared_ptr<Client> client;
-};
-
-struct AuthContext
-{
-    std::string token;
-    server::TokenPayload token_payload;
-};
+namespace tests {
 
 void
 userSignUp(TestEnvironment const& env)
@@ -311,7 +194,7 @@ FullTest::onRun()
 
     std::jthread server_thread{ [&server] { server.run(); } };
 
-    allTests();
+    tests::allTests();
 
     server.stop();
     db_conn_provider->stop();
